Fixed refit.dll not being included in BGservice artifacts
Refit was a dependency that Visual Studio did not realize was
a package dependency (not used directly). As such, it was left
out of the build folder.

I deleted all dependencies, and re-added Notakey.SDK (in the process
upgrading from 1.0.1.7 to 1.1.41) and Visual Studio correctly
identified Refit as a dependency.